defining state in state.py
